backend/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
daphne==4.1.2
deepface==0.0.75
Django==5.0.7
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
gdown==5.2.0
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
Pillow==10.4.0
protobuf==5.27.2
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
retina-face==0.0.13
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
psycopg2-binary==2.9.9
dj-database-url==2.2.0
tf-keras==2.15.1
librosa==0.10.2

# AI Psychologist Dependencies
PyJWT==2.8.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.32.3
django-cors-headers==4.3.1
